samples/vertexai/image/run/python/requirements.txt
Flask==2.1.0
gunicorn==20.1.0
google-auth
google-cloud-storage
google-cloud-aiplatform
Pillow
requests